+++ cmd/block-metrics/pkg/store.go
@@ -26,9 +26,9 @@ type Block struct {
 	FeeFrac uint64    `json:"fee_frac"`
 }
 
-func (s *Store) RegisterDomain(ctx context.Context, msg *types.MsgRegisterDomain) (int64, error) {
+func (st *Store) RegisterDomain(ctx context.Context, msg *types.MsgRegisterDomain) (int64, error) {
 	var id int64
-	err := s.db.QueryRowContext(ctx, `
+	err := st.db.QueryRowContext(ctx, `
 		INSERT INTO domains (name, admin, type, broker, fee_payer_addr)
 		VALUES ($1, $2, $3, $4, $5)
 		RETURNING id
@@ -36,33 +36,33 @@ func (s *Store) RegisterDomain(ctx context.Context, msg *types.MsgRegisterDomain
 	return id, castPgErr(err)
 }
 
-func (s *Store) DeleteDomain(ctx context.Context, msg *types.MsgDeleteDomain) error {
+func (st *Store) DeleteDomain(ctx context.Context, msg *types.MsgDeleteDomain) error {
 	sqlStatement := `
 		UPDATE domains SET deleted_at = now() 
 		WHERE name = $1`
-	_, err := s.db.ExecContext(ctx, sqlStatement, msg.Domain)
+	_, err := st.db.ExecContext(ctx, sqlStatement, msg.Domain)
 	return err
 }
 
-func (s *Store) TransferDomain(ctx context.Context, msg *types.MsgTransferDomain) error {
+func (st *Store) TransferDomain(ctx context.Context, msg *types.MsgTransferDomain) error {
 	sqlStatement := `
 	UPDATE domains
 	SET admin = $1
 	WHERE name = $2`
-	_, err := s.db.ExecContext(ctx, sqlStatement, msg.NewAdmin, msg.Domain)
+	_, err := st.db.ExecContext(ctx, sqlStatement, msg.NewAdmin, msg.Domain)
 	return err
 }
 
-func (s *Store) RegisterAccount(ctx context.Context, msg *types.MsgRegisterAccount) (int64, error) {
+func (st *Store) RegisterAccount(ctx context.Context, msg *types.MsgRegisterAccount) (int64, error) {
 	var accountID int64
 
 	sqlStatement := `SELECT id FROM domains WHERE name = $1;`
-	row := s.db.QueryRowContext(ctx, sqlStatement, msg.Domain)
+	row := st.db.QueryRowContext(ctx, sqlStatement, msg.Domain)
 	var domainID int64
 	if err := row.Scan(&domainID); err != nil {
 		return domainID, castPgErr(err)
 	}
-	_, err := s.db.ExecContext(ctx, `
+	_, err := st.db.ExecContext(ctx, `
 		INSERT INTO accounts (domain_id, domain, name, owner, registerer, broker, fee_payer_addr)
 		VALUES ($1, $2, $3, $4, $5, $6, $7)
 	`, domainID, msg.Domain, msg.Name, msg.Owner, msg.Registerer, msg.Broker, msg.FeePayerAddr)
@@ -72,17 +72,17 @@ func (s *Store) RegisterAccount(ctx context.Context, msg *types.MsgRegisterAccou
 	return accountID, nil
 }
 
-func (s *Store) ReplaceAccountResources(ctx context.Context, msg *types.MsgReplaceAccountResources) (int64, error) {
+func (st *Store) ReplaceAccountResources(ctx context.Context, msg *types.MsgReplaceAccountResources) (int64, error) {
 	var resourceID int64
 
 	sqlStatement := `SELECT id FROM accounts WHERE domain = $1 and name = $2;`
-	row := s.db.QueryRowContext(ctx, sqlStatement, msg.Domain, msg.Name)
+	row := st.db.QueryRowContext(ctx, sqlStatement, msg.Domain, msg.Name)
 	var accountID int64
 	if err := row.Scan(&accountID); err != nil {
 		return accountID, castPgErr(err)
 	}
 
-	tx, err := s.db.Begin()
+	tx, err := st.db.Begin()
 	if err != nil {
 		return accountID, castPgErr(err)
 	}
@@ -112,16 +112,16 @@ func (s *Store) ReplaceAccountResources(ctx context.Context, msg *types.MsgRepla
 	return resourceID, castPgErr(err)
 }
 
-func (s *Store) DeleteAccount(ctx context.Context, msg *types.MsgDeleteAccount) error {
+func (st *Store) DeleteAccount(ctx context.Context, msg *types.MsgDeleteAccount) error {
 	sqlStatement := `
 		UPDATE accounts SET deleted_at = now() 
 		WHERE domain = $1 AND name = $2`
-	_, err := s.db.ExecContext(ctx, sqlStatement, msg.Domain, msg.Name)
+	_, err := st.db.ExecContext(ctx, sqlStatement, msg.Domain, msg.Name)
 	return err
 }
 
-func (s *Store) InsertBlock(ctx context.Context, b Block) error {
-	tx, err := s.db.BeginTx(ctx, nil)
+func (st *Store) InsertBlock(ctx context.Context, b Block) error {
+	tx, err := st.db.BeginTx(ctx, nil)
 	if err != nil {
 		return errors.Wrap(err, "cannot create transaction")
 	}
@@ -141,15 +141,15 @@ func (s *Store) InsertBlock(ctx context.Context, b Block) error {
 	return wrapPgErr(err, "commit block tx")
 }
 
-func (s *Store) LatestBlock(ctx context.Context) (*Block, error) {
-	blocks, err := s.LastNBlock(ctx, 1, 0)
+func (st *Store) LatestBlock(ctx context.Context) (*Block, error) {
+	blocks, err := st.LastNBlock(ctx, 1, 0)
 	if err != nil {
 		return nil, err
 	}
 	return blocks[0], nil
 }
 
-func (s *Store) LastNBlock(ctx context.Context, limit, after int) ([]*Block, error) {
+func (st *Store) LastNBlock(ctx context.Context, limit, after int) ([]*Block, error) {
 	// max number of blocks that is allowed to retrieved is 100
 	if limit > 100 {
 		return nil, errors.Wrapf(ErrLimit, "limit exceeded")
@@ -158,14 +158,14 @@ func (s *Store) LastNBlock(ctx context.Context, limit, after int) ([]*Block, err
 	var rows *sql.Rows
 	var err error
 	if after == 0 {
-		rows, err = s.db.QueryContext(ctx, `
+		rows, err = st.db.QueryContext(ctx, `
 		SELECT block_height, block_hash, block_time, fee_frac
 		FROM blocks
 		ORDER BY block_height DESC
 		LIMIT $1
 	`, limit)
 	} else {
-		rows, err = s.db.QueryContext(ctx, `
+		rows, err = st.db.QueryContext(ctx, `
 		SELECT block_height, block_hash, block_time, fee_frac
 		FROM blocks
 		WHERE block_height < $1
@@ -204,4 +204,35 @@ func (s *Store) LastNBlock(ctx context.Context, limit, after int) ([]*Block, err
 		return nil, errors.Wrap(ErrNotFound, "no blocks")
 	}
 	return blocks, nil
+}
+func (st *Store) InsertGenesis(ctx context.Context, tmc *TendermintClient) error {
+	gen, err := FetchGenesis(ctx, tmc)
+	if err != nil {
+		return errors.Wrapf(err, "genesis fetch failed")
+	}
+	for _, domain := range gen.Domains {
+		msg := types.MsgRegisterDomain{
+			Name:         domain.Name,
+			Admin:        domain.Admin,
+			DomainType:   domain.Type,
+			Broker:       domain.Broker,
+			FeePayerAddr: domain.Admin,
+		}
+		if _, err := st.RegisterDomain(ctx, &msg); err != nil {
+			return errors.Wrapf(err, "cannot insert domain")
+		}
+	}
+	for _, acc := range gen.Accounts {
+		msg := types.MsgRegisterAccount{
+			Domain:    acc.Domain,
+			Name:      acc.Name,
+			Owner:     acc.Owner,
+			Resources: acc.Resources,
+			Broker:    acc.Broker,
+		}
+		if _, err := st.RegisterAccount(ctx, &msg); err != nil {
+			return errors.Wrapf(err, "cannot insert domain")
+		}
+	}
+	return err
 }